LayoutTests/http/tests/security/xss-DENIED-frame-name-expected.txt

CONSOLE MESSAGE: line 2: SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
   This test passes if it doesn't alert fail.
